Specify development dependencies in Gemfile.
Turns out that rubygems >= 2.2.0 doesn't allow duplicate dependencies in the gemspec.  I'm not sure if this is intentional behaviour when the "duplicates" are runtime vs. development dependencies (raised as https://github.com/rubygems/rubygems/issues/1032), but even if it is wrong, we can't release the gem until a fix comes out.

We can just remove the development dependencies from the specification and define them in the project gemfile instead.  It's not as clear a signal, but it'll do for now.