Add note about NameError: uninitialized constant RSpec::Core
The solution seems to be not requiring the gem until after rspec and require: false in the gemfile is the easiest way to manage that.  It may not be the best way, but it does fix it.  Probably want to investigate why and see if we can tidy it up later.